+++ docs/src/GroupAvatar.doc.js
@@ -30,6 +30,12 @@ card(
         required: true,
       },
       {
+        name: 'outline',
+        type: 'boolean',
+        defaultValue: false,
+        description: 'Adds a white border around GroupAvatar',
+      },
+      {
         name: 'size',
         type: `"sm" | "md" | "lg"`,
         description: 'sm: 24px, md: 40px, lg: 72px',